fix: address architecture review errors for PR #116
- Move sphobjinv and defusedxml from runtime to optional build deps,
  make imports lazy in manifest_builder.py (E1)
- Extract clean_rtd_markdown to Foundation text_utils.py, fixing the
  layer violation where manifest_builder imported from Domain (E3)
- Add _PERMALINK_RE to strip MkDocs/Sphinx permalink icons from
  extracted titles (E2)
- Return [] from _search_rtd_api when source has no RTD mapping
  instead of falling through to search all projects (E4)
- Add encoding="utf-8" to _load_manifest_file open() call (E5)
- Extract resp data into locals inside try block before finally
  closes the HTTP client in fetch_doc_content (E6)
