+++ pyproject.toml
@@ -31,8 +31,6 @@ dependencies = [
     "jinja2>=3.1",
     "pyyaml>=6.0",
     "typing_extensions>=4.0; python_version < '3.12'",
-    "sphobjinv>=2.3",
-    "defusedxml>=0.7",
 ]
 
 [project.urls]
@@ -45,6 +43,7 @@ ansible-know-mcp = "ansible_know.server:main"
 
 [project.optional-dependencies]
 dev = ["pytest>=7.0", "pytest-asyncio>=0.23", "pytest-cov>=4.0", "ruff>=0.4"]
+build = ["sphobjinv>=2.3", "defusedxml>=0.7"]
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/ansible_know"]
